extracted select index handler
