+++ handlers/input_handlers.py
@@ -115,64 +115,6 @@ class AskUserEventHandler(EventHandler):
         return MainGameEventHandler(self.engine)
 
 
-class SelectIndexHandler(AskUserEventHandler):
-    """Handles asking the user for an index on the map."""
-
-    def __init__(self, engine: Engine):
-        """Sets the cursor to the player when this handler is constructed."""
-        super().__init__(engine)
-        player = self.engine.player
-        engine.mouse_location = player.x, player.y
-
-    def on_render(self, console: tcod.console.Console) -> None:
-        """Highlight the tile under the cursor."""
-        super().on_render(console)
-
-        x, y = self.engine.mouse_location
-        console.rgb["bg"][x, y] = color.white
-        console.rgb["fg"][x, y] = color.black
-
-    def on_index_selection_changed(self) -> None:
-        pass
-
-    def ev_keydown(self, event: tcod.event.KeyDown) -> Optional[ActionOrHandler]:
-        """Check for key movement or confirmation keys."""
-        key = event.sym
-        if key in keys.MOVE_KEYS:
-            modifier = 1  # Holding modifier keys will speed up key movement.
-            if event.mod & (tcod.event.KMOD_LSHIFT | tcod.event.KMOD_RSHIFT):
-                modifier *= 5
-            if event.mod & (tcod.event.KMOD_LCTRL | tcod.event.KMOD_RCTRL):
-                modifier *= 10
-            if event.mod & (tcod.event.KMOD_LALT | tcod.event.KMOD_RALT):
-                modifier *= 20
-
-            x, y = self.engine.mouse_location
-            dx, dy = keys.MOVE_KEYS[key]
-            x += dx * modifier
-            y += dy * modifier
-            # Clamp the cursor index to the map size.
-            x = max(0, min(x, self.engine.game_map.width - 1))
-            y = max(0, min(y, self.engine.game_map.height - 1))
-            self.engine.mouse_location = x, y
-            self.on_index_selection_changed()
-            return None
-        elif key in keys.CONFIRM_KEYS:
-            return self.on_index_selected(*self.engine.mouse_location)
-        return super().ev_keydown(event)
-
-    def ev_mousebuttondown(self, event: tcod.event.MouseButtonDown) -> Optional[ActionOrHandler]:
-        """Left click confirms a selection."""
-        if self.engine.game_map.in_bounds(*event.tile):
-            if event.button == 1:
-                return self.on_index_selected(*event.tile)
-        return super().ev_mousebuttondown(event)
-
-    def on_index_selected(self, x: int, y: int) -> Optional[ActionOrHandler]:
-        """Called when an index is selected."""
-        raise NotImplementedError()
-
-
 class MainGameEventHandler(EventHandler):
     def ev_keydown(self, event: tcod.event.KeyDown) -> Optional[ActionOrHandler]:
         key = event.sym
