+++ handlers/single_ranged_attack_handler.py
@@ -1,6 +1,6 @@
 from actions.action import Action
 from engine import Engine
-from handlers.input_handlers import SelectIndexHandler
+from handlers.select_index_handler import SelectIndexHandler
 from typing import Optional
